Refactor and re-organize extension methods
